requirements.txt
hikari>=2.0.0.dev110
hikari-lightbulb>=2.3.1
python-dotenv>=0.20.0
aiohttp>=3.8.1